% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssocTest-function.R
\name{calc.skew}
\alias{calc.skew}
\title{calc.skew calculates the skew for a vector}
\usage{
calc.skew(x)
}
\arguments{
\item{x}{the vector whose skew is to be calcualted}
}
\value{
the skew
}
\description{
calc.skew calculates the skew for a vector
}
