% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-measure.R
\name{calcKer.IBS}
\alias{calcKer.IBS}
\title{calcKer.IBS calculate similarity matrix for genetic data}
\usage{
calcKer.IBS(X)
}
\arguments{
\item{X}{genetic data, with each row representing a subject}
}
\value{
a similarity matrix
}
\description{
calcKer.IBS calculate similarity matrix for genetic data
}
\examples{
library(MASS)
n = 100
k = 10
X = mvrnorm(n, rep(0,k), diag(k))
calcKer.IBS(X)
}
