% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-measure.R
\name{calcKerMat.gau}
\alias{calcKerMat.gau}
\title{calcKerMat.gau calculate similarity matrix via Gaussian kernel function}
\usage{
calcKerMat.gau(X, r = NULL)
}
\arguments{
\item{X}{the data to calculate its similarity matrix, with each row representing
a subject}

\item{r}{the scale parameter used in Gaussian kernel}
}
\value{
the centered similarity matrix
}
\description{
calcKerMat.gau calculate similarity matrix via Gaussian kernel function
}
\examples{
library(MASS)
n = 100
k = 10
x = mvrnorm(n, rep(0,k), diag(k))
calcKerMat.gau(x)
}
