% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-measure.R
\name{calcKerMat.poly}
\alias{calcKerMat.poly}
\title{calcKerMat.poly calculate similarity matrix via polynomial kernel function}
\usage{
calcKerMat.poly(X, r = 0, d = 5)
}
\arguments{
\item{X}{the data to calculate its similarity matrix, with each row representing
a subject}

\item{r}{the offset value uesed in kernel function}

\item{d}{the order of the kernel function}
}
\value{
the centered similarity matrix
}
\description{
calcKerMat.poly calculate similarity matrix via polynomial kernel function
}
\examples{
library(MASS)
n = 100
k = 10
x = mvrnorm(n, rep(0,k), diag(k))
calcKerMat.poly(x)
}
