% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-measure.R
\name{geneSimi}
\alias{geneSimi}
\title{geneSimi calculate similarity matrix for genetic data}
\usage{
geneSimi(x)
}
\arguments{
\item{x}{genetic data, with each column representing a subject}
}
\value{
a similarity matrix
}
\description{
geneSimi calculate similarity matrix for genetic data
}
\examples{
library(MASS)
n = 100
k = 10
x = mvrnorm(n, rep(0,k), diag(k))
geneSimi(t(x))
}
