% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssocTest-function.R
\name{KAT}
\alias{KAT}
\title{KAT Kernel based association test}
\usage{
KAT(E, G, B1 = 1000)
}
\arguments{
\item{E}{the similarity matrix for genetic data}

\item{G}{the similarity matrix for phenotypic data}

\item{B1}{permutation time}
}
\value{
the p value
}
\description{
KAT Kernel based association test
}
\examples{
library(MASS)
n = 100
p = 10
k = 10
betax = matrix(rnorm(k*p,0,1),k,p)
x = mvrnorm(n, rep(0,k), diag(k))
y = x\%*\%betax + mvrnorm(n, rep(0,p), diag(p))
G = calcKerMat.gau(x)
E = calcKerMat.gau(y)
KAT(E,G,100)
}
