% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssocTest-function.R
\name{MaxKAT}
\alias{MaxKAT}
\title{MaxKAT A maximum kernel-based association test to detect the
pleiotropic genetic effects on multiple phenotypes}
\usage{
MaxKAT(E, G, kVec, mVec, numPer = 1000, ifGEV = TRUE)
}
\arguments{
\item{E}{the similarity matrix for genetic data}

\item{G}{the similarity matrix for phenotypic data}

\item{kVec}{the index set for E}

\item{mVec}{the index set for G}

\item{numPer}{the number for permutation}

\item{ifGEV}{the method used for significance calculation, must be one
"NULL","TRUE" and "Both". If ifGEV = NULL, the p value based on permutation
is returned. If ifGEV = TRUE, the p value based on GEV approximation
is returned. If ifGEV = "Both", both p values based on permutation and GVE
approximation is returned.}
}
\value{
the p value for MaxKAT
}
\description{
MaxKAT A maximum kernel-based association test to detect the
pleiotropic genetic effects on multiple phenotypes
}
\examples{
library(MASS)
library(ismev)
library(extRemes)
n = 100
p = 10
k = 10
betax = matrix(rnorm(k*p,0,1),k,p)
x = mvrnorm(n, rep(0,k), diag(k))
y = x\%*\%betax + mvrnorm(n, rep(0,p), diag(p))
G = calcKerMat.gau(x)
E = calcKerMat.gau(y)
kVec = c(1/2,3/4,1,2,3)
mVec = c(1/2,3/4,1,2,3)
MaxKAT(E,G,kVec,mVec)
}
