% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-measure.R
\name{modify.normalization}
\alias{modify.normalization}
\title{modify.normalization data normalization}
\usage{
modify.normalization(x)
}
\arguments{
\item{x}{the data to be normalized}
}
\value{
the normalized data
}
\description{
modify.normalization data normalization
}
\examples{
library(MASS)
n = 100
x = rnorm(n, 0, 1)
modify.normalization(x)
}
